Add basic support to run unittests in GitHub Actions.
Includes a .gitignore file so we do not add unwanted files accidentally.
A very basic unittest so the module can at least be loaded, and a starting point for more unittests.
Clarify the macOS support in the README.
Add more pypi classifier for the project.
Update the repository url to the new home.